fix(ci): resolve clippy -D warnings across workspace
Unblocks PR #16 CI after the master merge by fixing lint in core, server, CLI, finetuning, and bench targets.

+++ oxidize-core/src/compute/numa.rs
@@ -111,7 +111,7 @@ mod imp {
                 len,
                 2usize,
                 mask.as_ptr() as usize,
-                (words * 64) as usize,
+                words * 64,
                 0u32,
             );
             if r != 0 {
